Refactor intcode computer
In implementing part 1, I had assumed that all opcodes would operate
on the three elements following the opcode. Furthermore, I assumed
that the opcode would operate on the elements that could be found in
the positions defined by the two elements following the opcode.

In reading day 2part 2, it became apparent to me that these
assumptions may not hold for future opcodes. Furthermore, it does not
make sense for code 1 to share so much code with code 2. In the
unlikely event that I am expected to come back and modify the
operation for code 1 or code 2, there does not seem to be a reason for
code 1 and 2 to change together.